+++ src/lib/AuthPlugin/BasicAuthPlugin.js
@@ -1,0 +1,16 @@
+import AuthPluginInterface from 'AuthPluginInterface'
+
+/**
+ *
+ */
+export default class BasicAuthPlugin extends AuthPluginInterface {
+  async authenticate (session, options) {
+    return {
+      ...options,
+      auth: {
+        user: session.cas.user,
+        pass: session.cas.pass,
+      },
+    }
+  }
+}
